change lesson-6
