+++ index-chat.html
@@ -18,9 +18,9 @@
         socket.on('connect', () => {
             console.log('Connected Socket!');
         });
-        const messageBlock = document.querySelector('#messages');
-        const sendButton = document.querySelector('#button');
         const input = document.querySelector('#input');
+        const sendButton = document.querySelector('#button');
+        const messageBlock = document.querySelector('#messages');
 
         sendButton.addEventListener('click', () => {
             const data = {
@@ -31,20 +31,20 @@
             input.value = '';
         });
 
-        // событие сообщение от сервера
+        // подписываемся на событие сообщение от сервера
         socket.on('server-msg', payload => {
             messageBlock.insertAdjacentHTML('beforeend', `<div>userId: ${payload.id} message: ${payload.message}</div>`)
         });
 
-        // событие подключения нового пользователя
+        // подписываемся на событие подключения нового пользователя
         socket.on('newUser', payload => {
-            messageBlock.insertAdjacentHTML('beforeend', `<div>Подключился новый пользователь userId: ${payload.id}</div>`)
+            messageBlock.insertAdjacentHTML('beforeend', `<div>Подключился новый пользователь с Id: ${payload.id}</div>`)
             messageBlock.insertAdjacentHTML('beforeend', `<div>Количество пользователей: ${payload.clientsCount}</div>`)
         });
 
-        // событие отключения пользователя
+        // подписываемся на событие отключения пользователя
         socket.on('disconectUser', payload => {
-            messageBlock.insertAdjacentHTML('beforeend', `<div>Отключился пользователь userId: ${payload.id}</div>`)
+            messageBlock.insertAdjacentHTML('beforeend', `<div>Отключился пользователь с Id: ${payload.id}</div>`)
             messageBlock.insertAdjacentHTML('beforeend', `<div>Количество пользователей: ${payload.clientsCount}</div>`)
         });
