+++ package.json
@@ -4,7 +4,7 @@
   "description": "Lesson GU Node.js",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js",
+    "start": "node lesson-6.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
